requirements.txt
future==1.0.0
iso8601==2.1.0
psutil==5.9.8
pyserial==3.5
PyYAML==6.0.1

